requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pypdf
docx2txt
streamlit
python-dotenv
reportlab
